fix: separate module for `ApiOptionsContext`
