+++ packages/plugin/src/components/ApiOptionsContext.tsx
@@ -1,0 +1,7 @@
+/* eslint-disable react/jsx-filename-extension */
+import { createContext } from "react";
+
+export const ApiOptionsContext = createContext({
+	hideInherited: false,
+	setHideInherited: (hideInherited: boolean) => {},
+});
